RAI-990: Split performance module into projection + report submodules
Mechanical split (PR #820 review thread #16): the ~2500-line performance module is divided into the write-side reactor (projection.rs) and read-side report assembly (report.rs), with shared helpers kept in the parent. No behavior change; 104 hedge-latency tests unchanged.
